rm cse
