+++ datafusion/sqllogictest/test_files/aggregate.slt
@@ -5375,3 +5375,25 @@ select sum(*) from t1;
 
 statement ok
 drop table t1;
+
+
+# test common subexpression eliminate
+statement ok
+create table t1(a bigint, b bigint, c bigint) as values (1, 2, 3), (null, 4, 2), (null, null, 2);
+
+statement ok
+set datafusion.explain.logical_plan_only = true;
+
+query TT
+explain select sum(a * (1 - b)), sum(a * (1 - b) * (1 + c)) from t1;
+----
+logical_plan
+01)Aggregate: groupBy=[[]], aggr=[[SUM({t1.a * (Int64(1) - t1.b)|{Int64(1) - t1.b|{t1.b}|{Int64(1)}}|{t1.a}} AS t1.a * Int64(1) - t1.b), SUM({t1.a * (Int64(1) - t1.b)|{Int64(1) - t1.b|{t1.b}|{Int64(1)}}|{t1.a}} AS t1.a * Int64(1) - t1.b * (Int64(1) + t1.c))]]
+02)--Projection: t1.a * (Int64(1) - t1.b) AS {t1.a * (Int64(1) - t1.b)|{Int64(1) - t1.b|{t1.b}|{Int64(1)}}|{t1.a}}, t1.c
+03)----TableScan: t1 projection=[a, b, c]
+
+statement ok
+set datafusion.explain.logical_plan_only = false;
+
+statement ok
+drop table t1;
